Fix crash due to wrong list indexing
ubi.block.layout.associate_blocks() appends a list of blocks as the last
element of each "layout_pair" object. When the layout_pair has a number
of layout_blocks other than 2 the newly added element will not have index
"2" as it was assumed before this fix. Use the "-1" index to always refer
to the last element in the list.