feat: parse the 'hash' field after 'key' in IDX:branch block
This  PR introduces the ability to read the `hash` after the `key`.
Nothing is being done with that `hash` for now. The immediate goal
is to have a functional `extract_files` operation.

Quoting the kernel's `ubifs-media.h` header file:

    In an authenticated UBIFS we have the hash of the referenced node
    after @key. This can't be added to the struct type definition
    because @key is a dynamically sized element already.

The size of key+hash is computed dynamically by dividing the remaining
length of the node by the number of children. A possible improvement
could be to grab the actual hash algorithm (stored somewhere?) and
verify that the hash length is the valid one.

A future improvement could be to have an `--check-authentication`
optional argument to the UBIFS operation what would verify that the
UBIFS image has not been tampered with.